Parte 1 da implementação funcionando

+++ Trabalho 1/src/fibonacci_iterative.cpp
@@ -1,6 +1,6 @@
 #include "../include/defines.hpp"
 
-int fibonacci_iterative(int n)
+size_t fibonacci_iterative(int n)
 {
 
     int a = 0, b = 1;
